test: checkpoint bankruptcy t22 tests
