+++ clients/rust/marginfi-cli/Cargo.toml
@@ -52,3 +52,4 @@ chrono = "0.4.23"
 bincode = "1.3.1"
 bs58 = "0.4.0"
 solana-address-lookup-table-program = "1.14"
+rand = "0.8.4"
