+++ programs/marginfi/tests/admin_actions/create_marginfi_group.rs
@@ -1,0 +1,50 @@
+use anchor_lang::{InstructionData, ToAccountMetas};
+use fixtures::prelude::*;
+use marginfi::prelude::MarginfiGroup;
+use pretty_assertions::assert_eq;
+use solana_program::{instruction::Instruction, system_program};
+use solana_program_test::*;
+use solana_sdk::{signature::Keypair, signer::Signer, transaction::Transaction};
+
+#[tokio::test]
+async fn marginfi_group_create_success() -> anyhow::Result<()> {
+    let test_f = TestFixture::new(None).await;
+
+    // Create & initialize marginfi group
+    let marginfi_group_key = Keypair::new();
+
+    let accounts = marginfi::accounts::MarginfiGroupInitialize {
+        marginfi_group: marginfi_group_key.pubkey(),
+        admin: test_f.payer(),
+        system_program: system_program::id(),
+    };
+    let init_marginfi_group_ix = Instruction {
+        program_id: marginfi::id(),
+        accounts: accounts.to_account_metas(Some(true)),
+        data: marginfi::instruction::MarginfiGroupInitialize {}.data(),
+    };
+    let tx = Transaction::new_signed_with_payer(
+        &[init_marginfi_group_ix],
+        Some(&test_f.payer().clone()),
+        &[&test_f.payer_keypair(), &marginfi_group_key],
+        test_f.get_latest_blockhash().await,
+    );
+    let res = test_f
+        .context
+        .borrow_mut()
+        .banks_client
+        .process_transaction(tx)
+        .await;
+
+    assert!(res.is_ok());
+
+    // Fetch & deserialize marginfi group account
+    let marginfi_group: MarginfiGroup = test_f
+        .load_and_deserialize(&marginfi_group_key.pubkey())
+        .await;
+
+    // Check basic properties
+    assert_eq!(marginfi_group.admin, test_f.payer());
+
+    Ok(())
+}
