+++ programs/marginfi/tests/admin_actions/mod.rs
@@ -1,3 +1,7 @@
 mod account_transfer;
-mod marginfi_group;
+mod bankruptcy;
+mod bankruptcy_auth;
+mod create_marginfi_group;
+mod interest_accrual;
+mod setup_bank;
 mod withdraw_fees;
