+++ programs/marginfi/tests/misc/mod.rs
@@ -1,0 +1,7 @@
+mod bank_ignore_stale_isolated_banks;
+mod bank_variable_oracle_staleness;
+mod collateral_value_cap;
+mod operational_state;
+mod regression;
+mod risk_engine_flexible_oracle_checks;
+mod token_extensions;
