+++ programs/marginfi/tests/tests.rs
@@ -1,2 +1,3 @@
 mod admin_actions;
+mod misc;
 mod user_actions;
