+++ programs/marginfi/tests/user_actions/liquidate.rs
@@ -15,6 +15,8 @@ use test_case::test_case;
 #[test_case(100., 9.9, 1., BankMint::Usdc, BankMint::Sol)]
 #[test_case(123., 122., 10., BankMint::SolEquivalent, BankMint::SolEqIsolated)]
 #[test_case(1_000., 999., 10., BankMint::Usdc, BankMint::T22WithFee)]
+#[test_case(2_000., 99., 1_000., BankMint::T22WithFee, BankMint::SolEquivalent)]
+#[test_case(2_000., 1_999., 2_000., BankMint::Usdc, BankMint::PyUSD)]
 #[tokio::test]
 async fn marginfi_account_liquidation_success(
     deposit_amount: f64,
